#50, #51 - Изменена структура пакетов - Пакеты и классы названы в соответствии с #51 - Модуль валидации вынесен в соответствии с #50
