+++ src/main/scala/PhoneBook.scala
@@ -1,15 +1,20 @@
-package www
+import api.Api.indexCall
 
-import cats.effect._
+import cats.effect.{ContextShift, ExitCode, IO, IOApp}
 import cats.implicits._
-import org.http4s.server.Router
+
 import org.http4s.server.blaze.BlazeServerBuilder
-import org.http4s.implicits._
 import org.http4s.server.middleware.Logger
-//import org.http4s.server.middleware.{Logger, RequestLogger, ResponseLogger}
+import org.http4s.implicits._
+import org.http4s.server.Router
 
 import scala.concurrent.ExecutionContext.Implicits.global
-import Api._
+
+import api.CorsApiWrapper
+import api.Api._
+
+
+
 
 object PhoneBook extends IOApp{
   implicit val cs: ContextShift[IO] = IO.contextShift(global)
