+++ src/main/scala/share/InputValidator.scala
@@ -1,0 +1,9 @@
+package share
+
+object InputValidator {
+  def isNameValid(name: String): Boolean =
+    !name.isBlank
+
+  def isPhoneValid(number: String): Boolean =
+    !number.isBlank
+}
